correctly flip non-traceback alignments
This is a likely candidate for the resolution of #1039

This makes it impossible for non-first best alignments to be promoted to the
alignment that is output.

It does also seem to matter for pair rescue, but pair rescue doesn't seem to
help us in many contexts.

+++ src/mapper.hpp
@@ -290,7 +290,7 @@ public:
     void record_fragment_configuration(int length, const Alignment& aln1, const Alignment& aln2);
     
     string fragment_model_str(void);
-    void save_frag_lens_to_alns(Alignment& aln1, Alignment& aln2, const map<string, int>& approx_frag_lengths, bool is_consistent);
+    void save_frag_lens_to_alns(Alignment& aln1, Alignment& aln2, const map<string, int64_t>& approx_frag_lengths, bool is_consistent);
     
     // These functions are the authorities on the estimated parameters
     double fragment_length_stdev(void);
@@ -307,19 +307,19 @@ public:
     bool cached_fragment_direction = 1;
     
     // These variables are used to manage the periodic updates
-    int since_last_fragment_length_estimate = 0;
-    int fragment_model_update_interval = 10;
+    int64_t since_last_fragment_length_estimate = 0;
+    int64_t fragment_model_update_interval = 10;
     
     // These deques are used for the periodic running estimation of the fragment length distribution
     deque<double> fragment_lengths;
     deque<bool> fragment_orientations;
     deque<bool> fragment_directions;
 
-    int fragment_max = 10000; // the maximum length fragment which we will consider when estimating fragment lengths
-    int fragment_size = 0; // Used to bound clustering of MEMs during paired end mapping, also acts as sentinel to determine
+    int64_t fragment_max = 10000; // the maximum length fragment which we will consider when estimating fragment lengths
+    int64_t fragment_size = 0; // Used to bound clustering of MEMs during paired end mapping, also acts as sentinel to determine
                        // if consistent pairs should be reported; dynamically estimated at runtime
     double fragment_sigma = 10; // the number of times the standard deviation above the mean to set the fragment_size
-    int fragment_length_cache_size = 1000;
+    int64_t fragment_length_cache_size = 1000;
     float perfect_pair_identity_threshold = 0.95;
     bool fixed_fragment_model = true;
     
@@ -524,7 +524,7 @@ public:
     vector<Alignment> mem_to_alignments(MaximalExactMatch& mem);
 
     // fargment length estimation
-    map<string, int> approx_pair_fragment_length(const Alignment& aln1, const Alignment& aln2);
+    map<string, int64_t> approx_pair_fragment_length(const Alignment& aln1, const Alignment& aln2);
     int64_t first_approx_pair_fragment_length(const Alignment& aln1, const Alignment& aln2);
     // uses the cached information about the graph in the xg index to get an approximate node length
     double average_node_length(void);
